Update charter.md
I made a few spelling and grammar fixes. 

I replaced TSC with ZLC on line 94 assuming that is what you meant as we are not yet going to use OMP/LF TSC for resolutions.  This is in disagreement with comments by John Mertic in terms of roll-up.

A general note that I did not attempt to address in my edits is that the manifesto seems a departure in theme from earlier discussions on the TSC in that earlier we emphasized that technical and design roles of TSC were to be focused on those that were cross-squad in their nature.  This cross-squad emphasis seems now absent from manifesto and the opening line as well as item 3 of manifesto bullets are possibly to vague and expansive (does "technical direction" need to be qualified?).  We do not want the TSC to take over all technical direction of all Zowe components.

Signed-off-by: Peter Fandel pfandel@rocketsoftware.com